+++ src/Falco.Markup/TestHelpers.fs
@@ -121,6 +121,7 @@ module TestHelpers =
 
                                     | ParentNode _
                                     | SelfClosingNode _
+                                    | NodeList _
                                     | TextNode _ -> ()
 
                             // <option></option>
@@ -129,6 +130,7 @@ module TestHelpers =
 
                             | ParentNode _
                             | SelfClosingNode _
+                            | NodeList _
                             | TextNode _ -> () ]
 
                     if selected.Length = 0 then
@@ -141,6 +143,9 @@ module TestHelpers =
                     for c in children do
                         buildNameValues c
 
+                | NodeList nodes ->
+                    for node in nodes do
+                        buildNameValues node
                 | SelfClosingNode _
                 | TextNode _ -> ()